+++ src/test/samples/shellscript.sh
@@ -1,0 +1,6 @@
+# ! This is a comment!
+echo Hello World	# This is a comment, too!
+
+# Set logfile
+# shellcheck disable=SC2034
+logfile="$logDir/${0##*/}.log"