+++ src/parser.ts
@@ -83,7 +83,7 @@ export class Parser {
 	}
 
 	/**
-	 * Finds all single line comments delimted by a given delimter and matching tags specified in package.json
+	 * Finds all single line comments delimited by a given delimiter and matching tags specified in package.json
 	 * @param activeEditor The active text editor containing the code document
 	 */
 	public FindSingleLineComments(activeEditor: vscode.TextEditor): void {
@@ -128,7 +128,7 @@ export class Parser {
 		
 		let text = activeEditor.document.getText();
 
-		// Build up regex matcher for custom delimter tags
+		// Build up regex matcher for custom delimiter tags
 		let characters: Array<string> = [];
 		for (let commentTag of this.tags) {
 			characters.push(commentTag.escapedTag);
@@ -183,7 +183,7 @@ export class Parser {
 
 		let text = activeEditor.document.getText();
 
-		// Build up regex matcher for custom delimter tags
+		// Build up regex matcher for custom delimiter tags
 		let characters: Array<string> = [];
 		for (let commentTag of this.tags) {
 			characters.push(commentTag.escapedTag);
@@ -345,8 +345,9 @@ export class Parser {
 				this.setCommentFormat("--", "{-", "-}");
 				break;
 
-			case "vb":
+			case "brightscript":
 			case "diagram": // ? PlantUML is recognized as Diagram (diagram)
+			case "vb":
 				this.delimiter = "'";
 				break;
 
@@ -382,6 +383,7 @@ export class Parser {
 			
 			case "html":
 			case "markdown":
+			case "xml":
 				this.setCommentFormat("<!--", "<!--", "-->");
 				break;