+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Features
 * Adding Bold, Italic, and Underline ([e41ccc4](https://github.com/aaron-bond/better-comments/commit/e41ccc4)), closes [#50](https://github.com/aaron-bond/better-comments/issues/50). Massive thanks to _Fr33maan_
 
+### Bug Fixes
+* Fixing Shaderlab support ([]()), closes [#245](https://github.com/aaron-bond/better-comments/issues/245)
+
 ## [2.0.5] (2019-05-15)
 ### Features
 * Adding Markdown support ([54e51fb](https://github.com/aaron-bond/better-comments/commit/54e51fb)), closes [#91](https://github.com/aaron-bond/better-comments/issues/91)