+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug Fixes
 * Fixing Shaderlab support ([]()), closes [#245](https://github.com/aaron-bond/better-comments/issues/245)
+* Fixing SASS support ([]()), closes [#123](https://github.com/aaron-bond/better-comments/issues/123), [#215](ttps://github.com/aaron-bond/better-comments/issues/215)
 
 ## [2.0.5] (2019-05-15)
 ### Features