+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [2.1.0] (????-??-??)
+## [2.1.0] (2020-07-13)
 ### Features
 * Adding Bold, Italic, and Underline ([e41ccc4](https://github.com/aaron-bond/better-comments/commit/e41ccc4)), closes [#50](https://github.com/aaron-bond/better-comments/issues/50). Massive thanks to _Fr33maan_
 * Adding XML support